examples: cleanup Blinker example code for better readability
